% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequality_w.R
\name{Entropy}
\alias{Entropy}
\title{Compute entropy measure of inequality}
\usage{
Entropy(distri, weights, theta)
}
\arguments{
\item{distri}{Distribution of outcomes.}

\item{weights}{Weights used in the econometric model.}

\item{theta}{Value of sensibility.}
}
\value{
A value of inequality.
}
\description{
Compute entropy measure of inequality
}
