% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{exData}
\alias{exData}
\title{Example of a dataset with econometric variables.}
\format{
A data.frame with 1000 rows and 8 columns.
\describe{
\item{FULLTIME}{Whether the individual works a full time or not.}
\item{SEXE}{The sex of the individual.}
\item{AGEQ}{The age of the individual.}
\item{diplom}{The type of diploma.}
\item{salred}{The revenue.}
\item{POT_EXP}{The professional experience.}
\item{extridf}{xxxx}
\item{EXTRIDF_N}{xxxx}
}
}
\usage{
exData
}
\description{
This is a sample dataset to be used in the package examples. In this example,
we used data from xxxxx.
}
\keyword{datasets}
