% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{exTobitModel}
\alias{exTobitModel}
\title{Example of an econometric model.}
\format{
A model selection xxxx
\describe{
}
}
\usage{
exTobitModel
}
\description{
This is a toy model to be used in the package examples. In this example,
we used data from xxxxx.
}
\keyword{datasets}
