% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getCoalitions.R
\name{getCoalitions}
\alias{getCoalitions}
\title{Get all possible coalitions for a list of given factors in an econometric
model}
\usage{
getCoalitions(factors_list)
}
\arguments{
\item{factors_list}{A list of factors in the econometric model.}
}
\value{
A data.frame with factors in columns and all possible coalitions as
rows. The last column corresponds to the name of the coalition which is
defined based on attributes in or out of the coalition.
}
\description{
Get all possible coalitions for a list of given factors in an econometric
model
}
\examples{
factors <- list(A = "A", B = "B", CD = c("C", "D"), E = "E")
getCoalitions(factors_list = factors)
}
