% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getFactorList.R
\name{getFactorList}
\alias{getFactorList}
\title{Extract the list of factors from an econometric model}
\usage{
getFactorList(model_eco, database, residuals = FALSE, equation)
}
\arguments{
\item{model_eco}{An econometric model.}

\item{database}{A data.frame with data used to calibrate the model.}

\item{residuals}{Whether residuals are considered or not.}

\item{equation}{Either the "selection" equation or the "outcome" equation from a Tobit II model.}
}
\value{
A list of factors.
}
\description{
Extract the list of factors from an econometric model
}
\examples{
getFactorList(model_eco = exTobitModel,
              database = exData,
              residuals = TRUE,
              equation = "outcome")
}
