% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getInequality.R
\name{getInequality}
\alias{getInequality}
\title{Compute the inequality for a given coalition and a given econometric model.}
\usage{
getInequality(
  coalition,
  factors_list,
  model_eco,
  equation,
  measure = Atkinson,
  database,
  transfo = NULL,
  mXOutcome,
  correction = NA,
  errors,
  equaGame = FALSE,
  theta = NULL,
  weights
)
}
\arguments{
\item{coalition}{The coalition used for the computation of the inequality.}

\item{factors_list}{A list of factors in the econemetric model.}

\item{model_eco}{An econometric model.}

\item{equation}{Either the "selection" equation or the "outcome" equation from a Tobit II model.}

\item{measure}{The function to use to compute the inequality, either Gini(),
Gini_G(), Entropy(), Atkinson(), Kolm(), Var().}

\item{database}{A data.frame with data used to calibrate the model.}

\item{transfo}{Specify the inverse of the function used to transform the data.}

\item{mXOutcome}{A data.frame with xxxxx}

\item{correction}{Whether the results need to be corrected or not (either NA,
"abs" or "rel").}

\item{errors}{A vector with the model residuals.}

\item{equaGame}{Whether the game is equalized or not (either TRUE or FALSE).}

\item{theta}{A value of sensibility for the computation of the inequality.}

\item{weights}{A vector with the weights used in the econometric model (if
there are none use NA).}
}
\value{
The value of inequality for the considered coalition.
}
\description{
Compute the inequality for a given coalition and a given econometric model.
}
\examples{
data(Mroz87)
exTobitModel <- sampleSelection::selection(lfp ~ age + I(age^2) + faminc + kids5 + educ,
    wage ~ exper + I(exper^2) + educ + city,
    data = Mroz87)
factors <- getFactorList(
  equation = "outcome",
  model_eco = exTobitModel,
  database = Mroz87)
coa <- getCoalitions(factors_list = factors)
distrib <- getShapleyDistrib(
  model_eco = exTobitModel,
  equation = "outcome",
  database = Mroz87)
getInequality(
  coalition = coa[1, ],
  factors_list = factors,
  model_eco = exTobitModel,
  equation = "outcome",
  measure = Gini_w,
  database = Mroz87,
  transfo = NULL,
  mXOutcome = distrib,
  correction = NA,
  errors = NA,
  equaGame = FALSE,
  theta = NULL,
  weights = rep(1, nrow(Mroz87)))
}
