% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getInteractionTerms.R
\name{getInteractionTerms}
\alias{getInteractionTerms}
\title{Extract the interaction terms from the marginal contributions of attributes in each coalition.}
\usage{
getInteractionTerms(factors_list, coalitions, marginalContrib)
}
\arguments{
\item{factors_list}{A list of factors in the econometric model.}

\item{coalitions}{A data.frame with all coalitions used for the computation of the inequality.}

\item{marginalContrib}{The marginal contributions of attributes in each coalition.}
}
\value{
A list with 1) the interactions terms, 2) the relative interaction terms.
}
\description{
Extract the interaction terms from the marginal contributions of attributes in each coalition.
}
\examples{
factors <- getFactorList(
 model = "outcome",
 model_eco = exTobitModel,
 database = exData,
 residuals = FALSE)
coa <- getCoalitions(factors_list = factors)
distrib <- getShapleyDistrib(
 model_eco = exTobitModel,
 model = "outcome",
 database = exData)
ineq <- vector()
for(i in 1:3){
 ineq[i] <- getInequality(
   coalition = coa[i, ],
   factors_list = factors,
   model_eco = exTobitModel,
   model = "outcome",
   measure = Atkinson,
   database = exData,
   transfo = exp,
   mXOutcome = distrib,
   correction = NA,
   errors = NA,
   equaGame = FALSE,
   theta = 1,
   weights = exData$extridf)
}
margContrib <- getMarginalContrib(
 inequality = ineq,
 coalitions = coa,
 nVar = length(factors))
getInteractionTerms(factors_list = factors,
                   coalitions = coa,
                   marginalContrib = margContrib)
}
