% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getMarginalContrib.R
\name{getMarginalContrib}
\alias{getMarginalContrib}
\title{Compute the marginal contributions of each attribute in all coalitions.}
\usage{
getMarginalContrib(inequality, coalitions, nVar, wMC = TRUE)
}
\arguments{
\item{inequality}{A vector containing all inequalities.}

\item{coalitions}{A data.frame with all coalitions used for the computation of the inequality.}

\item{nVar}{The number of factors.}

\item{wMC}{All marginal contributions are weighted according to the Shapley weights (default, wMC=T), but the user might want non-weighted marginal contributions.}
}
\value{
A data.frame with the marginal contributions and inequalities for all attributes.
}
\description{
Compute the marginal contributions of each attribute in all coalitions.
}
\examples{
data(Mroz87)
exTobitModel <- sampleSelection::selection(lfp ~ age + I(age^2) + faminc + kids5 + educ,
    wage ~ exper + I(exper^2) + educ + city,
    data = Mroz87)
factors <- getFactorList(
  equation = "outcome",
  model_eco = exTobitModel,
  database = Mroz87)
coa <- getCoalitions(factors_list = factors)
distrib <- getShapleyDistrib(
  model_eco = exTobitModel,
  equation = "outcome",
  database = Mroz87)
ineq <- vector()
for(i in 1:nrow(coa)){
 ineq[i] <- getInequality(
  coalition = coa[i, ],
  factors_list = factors,
  model_eco = exTobitModel,
  equation = "outcome",
  measure = Gini_w,
  database = Mroz87,
  transfo = NULL,
  mXOutcome = distrib,
  correction = NA,
  errors = NA,
  equaGame = FALSE,
  theta = NULL,
  weights = rep(1, nrow(Mroz87)))
}
getMarginalContrib(
  inequality = ineq,
  coalitions = coa,
  nVar = length(factors),
  wMC = TRUE)
}
