% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getShapleyVal.R
\name{getShapleyVal}
\alias{getShapleyVal}
\title{Compute the Shapley values from marginal contributions of attributes.}
\usage{
getShapleyVal(marginalContrib)
}
\arguments{
\item{marginalContrib}{A data.frame wit marginal contributions.}
}
\value{
A data.frame with Shapley value for each attribute and relative value.
}
\description{
Compute the Shapley values from marginal contributions of attributes.
}
\examples{
data(Mroz87)
exTobitModel <- sampleSelection::selection(lfp ~ age + I(age^2) + faminc + kids5 + educ,
    wage ~ exper + I(exper^2) + educ + city,
    data = Mroz87)
factors <- getFactorList(
  equation = "outcome",
  model_eco = exTobitModel,
  database = Mroz87)
coa <- getCoalitions(factors_list = factors)
distrib <- getShapleyDistrib(
  model_eco = exTobitModel,
  equation = "outcome",
  database = Mroz87)
ineq <- vector()
for(i in 1:nrow(coa)){
 ineq[i] <- getInequality(
  coalition = coa[i, ],
  factors_list = factors,
  model_eco = exTobitModel,
  equation = "outcome",
  measure = Gini_w,
  database = Mroz87,
  transfo = NULL,
  mXOutcome = distrib,
  correction = NA,
  errors = NA,
  equaGame = FALSE,
  theta = NULL,
  weights = rep(1, nrow(Mroz87)))
}
margContrib <- getMarginalContrib(
  inequality = ineq,
  coalitions = coa,
  nVar = length(factors))
getShapleyVal(marginalContrib = margContrib)
}
