% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequality_w.R
\name{Gini_w}
\alias{Gini_w}
\title{Compute weighted Gini's measure of inequality}
\usage{
Gini_w(distri, weights)
}
\arguments{
\item{distri}{Distribution of outcomes.}

\item{weights}{Weights used in the econometric model.}
}
\value{
A value of inequality.
}
\description{
Compute weighted Gini's measure of inequality
}
\examples{
distri <- rnorm(mean = 20, n = 100)
weights <- rep(1, 100)
Gini_w(distri, weights)
}
