% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapley_Decompo_TOBIT.R
\name{shapleyDecompo}
\alias{shapleyDecompo}
\title{Compute the Shapley decomposition for all coalitions possible in an econometric model.}
\usage{
shapleyDecompo(
  database,
  model_eco,
  model = "selection",
  equaGame = TRUE,
  correction = NA,
  transfo = NULL,
  data_weights = rep(1, nrow(database)),
  residuals = FALSE,
  measure,
  theta = NULL
)
}
\arguments{
\item{database}{A data.frame with data used to calibrate the model.}

\item{model_eco}{An econometric model.}

\item{model}{The type of model (either "selection" or "outcome").}

\item{equaGame}{Whether the game is equalized or not (either TRUE or FALSE).}

\item{correction}{Whether the results need to be corrected or not (either NA, "abs" or "rel").}

\item{transfo}{Specify the inverse of the function used to transform the data (default is NULL).}

\item{data_weights}{A vector with the weights used in the econometric model (if there are none use NA).}

\item{residuals}{Whether the residuals of the model are considered in the decomposition or not (TRUE or FALSE).}

\item{measure}{The function to use to compute the inequality, either Gini(),
Gini_G(), Entropy(), Atkinson(), Kolm(), Var().}

\item{theta}{A value of sensibility for the computation of the inequality.}
}
\value{
Returns a list with: 1) The Shapley for each attribute, 2) The marginal contributions of each attributes, 3) The interaction terms.
}
\description{
Compute the Shapley decomposition for all coalitions possible in an econometric model.
}
\examples{
shapleyDecompo(database = exData,
               model_eco = exTobitModel,
               model = "outcome",
               equaGame = TRUE,
               correction = NA,
               data_weights = exData$extridf,
               residuals = TRUE,
               transfo = exp,
               measure = Atkinson,
               theta = 1)
}
