% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequality_w.R
\name{Var}
\alias{Var}
\title{Var}
\usage{
Var(distri, weights)
}
\arguments{
\item{distri}{Distribution of outcomes.}

\item{weights}{Weights used in the econometric model.}
}
\value{
A value of inequality.
}
\description{
Var
}
