% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequality_w.R
\name{Atkinson}
\alias{Atkinson}
\title{Atkinson}
\usage{
Atkinson(distri, weights, theta)
}
\arguments{
\item{distri}{Distribution of outcomes.}

\item{weights}{Weights used in the econometric model.}

\item{theta}{Value of sensibility.}
}
\value{
A value of inequality.
}
\description{
Atkinson
}
