% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getFactorList.R
\name{getFactorList}
\alias{getFactorList}
\title{Extract the list of factors from an econometric model}
\usage{
getFactorList(model_eco, database, residuals = FALSE, equation)
}
\arguments{
\item{model_eco}{An econometric model.}

\item{database}{A data.frame with data used to calibrate the model.}

\item{residuals}{Whether residuals are considered or not.}

\item{equation}{Either the "selection" equation or the "outcome" equation from a Tobit II model.}
}
\value{
The list of factors in the model.
}
\description{
Extract the list of factors from an econometric model
}
\examples{
data(Mroz87)
exTobitModel <- sampleSelection::selection(lfp ~ age + I(age^2) + faminc + kids5 + educ,
    wage ~ exper + I(exper^2) + educ + city,
    data = Mroz87)
getFactorList(model_eco = exTobitModel,
              database = Mroz87,
              residuals = TRUE,
              equation = "outcome")
}
