% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getInteractionTerms.R
\name{getInteractionTerms}
\alias{getInteractionTerms}
\title{Extract the interaction terms from the marginal contributions of attributes in each coalition.}
\usage{
getInteractionTerms(factors_list, coalitions, marginalContrib)
}
\arguments{
\item{factors_list}{A list of factors in the econometric model.}

\item{coalitions}{A data.frame with all coalitions used for the computation of the inequality.}

\item{marginalContrib}{The marginal contributions of attributes in each coalition.}
}
\value{
A list with 1) the interactions terms, 2) the relative interaction terms.
}
\description{
Extract the interaction terms from the marginal contributions of attributes in each coalition.
}
\examples{
data(Mroz87)
exTobitModel <- sampleSelection::selection(lfp ~ age + I(age^2) + faminc + kids5 + educ,
    wage ~ exper + I(exper^2) + educ + city,
    data = Mroz87)
factors <- getFactorList(
  equation = "outcome",
  model_eco = exTobitModel,
  database = Mroz87)
coa <- getCoalitions(factors_list = factors)
distrib <- getShapleyDistrib(
  model_eco = exTobitModel,
  equation = "outcome",
  database = Mroz87)
ineq <- vector()
for(i in 1:nrow(coa)){
 ineq[i] <- getInequality(
  coalition = coa[i, ],
  factors_list = factors,
  model_eco = exTobitModel,
  equation = "outcome",
  measure = Gini_w,
  database = Mroz87,
  transfo = NULL,
  mXOutcome = distrib,
  correction = NA,
  errors = NA,
  equaGame = FALSE,
  theta = NULL,
  weights = rep(1, nrow(Mroz87)))
}
margContrib <- getMarginalContrib(
inequality = ineq,
 coalitions = coa,
 nVar = length(factors))
getInteractionTerms(factors_list = factors,
                   coalitions = coa,
                   marginalContrib = margContrib)
}
