% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getMarginalContrib.R
\name{getMarginalContrib}
\alias{getMarginalContrib}
\title{Compute the marginal contributions of each attribute in all coalitions.}
\usage{
getMarginalContrib(inequality, coalitions, nVar)
}
\arguments{
\item{inequality}{A vector containing all inequalities.}

\item{coalitions}{A data.frame with all coalitions used for the computation of the inequality.}

\item{nVar}{The number of factors.}
}
\value{
A data.frame with the marginal contributions and inequalities for all attributes.
}
\description{
Compute the marginal contributions of each attribute in all coalitions.
}
\examples{
factors <- getFactorList(
 model = "outcome",
 model_eco = exTobitModel,
 database = exData,
 residuals = FALSE)
coa <- getCoalitions(factors_list = factors)
distrib <- getShapleyDistrib(
 model_eco = exTobitModel,
 model = "outcome",
 database = exData)
ineq <- vector()
for(i in 1:3){
 ineq[i] <- getInequality(
   coalition = coa[i, ],
   factors_list = factors,
   model_eco = exTobitModel,
   model = "outcome",
   measure = Atkinson,
   database = exData,
   transfo = exp,
   mXOutcome = distrib,
   correction = NA,
   errors = NA,
   equaGame = FALSE,
   theta = 1,
   weights = exData$extridf)
}
getMarginalContrib(
 inequality = ineq,
 coalitions = coa,
 nVar = length(factors))
}
