% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getShapleyDistrib.R
\name{getShapleyDistrib}
\alias{getShapleyDistrib}
\title{Obtain the distributions used in the computation of inequalities.}
\usage{
getShapleyDistrib(model_eco, database, equation)
}
\arguments{
\item{model_eco}{An econometric model.}

\item{database}{A data.frame with data used to calibrate the model.}

\item{equation}{Either the "selection" equation or the "outcome" equation from a Tobit II model.}
}
\value{
A list with 1) the distribution of xxx, 2) the distribution of xxx.
}
\description{
Obtain the distributions used in the computation of inequalities.
}
\examples{
data(Mroz87)
exTobitModel <- sampleSelection::selection(lfp ~ age + I(age^2) + faminc + kids5 + educ,
    wage ~ exper + I(exper^2) + educ + city,
    data = Mroz87)
factors <- getFactorList(model_eco = exTobitModel,
              database = Mroz87,
              residuals = TRUE,
              equation = "outcome")
coa <- getCoalitions(factors_list = factors)
getShapleyDistrib(
 model_eco = exTobitModel,
 equation = "outcome",
 database = Mroz87)
}
