% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getShapleyVal.R
\name{getShapleyVal}
\alias{getShapleyVal}
\title{Compute the Shapley values from marginal contributions of attributes.}
\usage{
getShapleyVal(marginalContrib)
}
\arguments{
\item{marginalContrib}{A data.frame wit marginal contributions.}
}
\value{
A data.frame with Shapley value for each attribute and relative value.
}
\description{
Compute the Shapley values from marginal contributions of attributes.
}
\examples{
factors <- getFactorList(
 model = "outcome",
 model_eco = exTobitModel,
 database = exData,
 residuals = FALSE)
coa <- getCoalitions(factors_list = factors)
distrib <- getShapleyDistrib(
 model_eco = exTobitModel,
 model = "outcome",
 database = exData)
ineq <- vector()
for(i in 1:3){
 ineq[i] <- getInequality(
   coalition = coa[i, ],
   factors_list = factors,
   model_eco = exTobitModel,
   model = "outcome",
   measure = Atkinson,
   database = exData,
   transfo = exp,
   mXOutcome = distrib,
   correction = NA,
   errors = NA,
   equaGame = FALSE,
   theta = 1,
   weights = exData$extridf)
}
margContrib <- getMarginalContrib(
 inequality = ineq,
 coalitions = coa,
 nVar = length(factors))
getShapleyVal(marginalContrib = margContrib)
}
