% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequality_w.R
\name{Gini_G}
\alias{Gini_G}
\title{Gini_G}
\usage{
Gini_G(distri, theta)
}
\arguments{
\item{distri}{Distribution of outcomes.}

\item{theta}{Value of sensibility.}

\item{weights}{Weights used in the econometric model.}
}
\value{
A value of inequality.
}
\description{
Gini_G
}
