% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequality_w.R
\name{Gini_w}
\alias{Gini_w}
\title{Gini_w}
\usage{
Gini_w(distri, weights)
}
\arguments{
\item{distri}{Distribution of outcomes.}

\item{weights}{Weights used in the econometric model.}
}
\value{
A value of inequality.
}
\description{
Gini_w
}
