% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequality_w.R
\name{Kolm}
\alias{Kolm}
\title{Compute Kolm's measure of inequality}
\usage{
Kolm(distri, weights, theta)
}
\arguments{
\item{distri}{Distribution of outcomes.}

\item{weights}{Weights used in the econometric model.}

\item{theta}{Value of sensibility.}
}
\value{
A value of inequality.
}
\description{
Compute Kolm's measure of inequality
}
\examples{
distri <- rnorm(mean = 20, n = 100)
weights <- rep(1, 100)
theta <- 1
Kolm(distri, weights, theta)
}
