% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_shapleyDecompo.R
\name{shapleyDecompo}
\alias{shapleyDecompo}
\title{Compute the Shapley decomposition of the inequality of an outcome variable across different attributes from an econometric model.}
\usage{
shapleyDecompo(
  database,
  model_eco,
  equation = "selection",
  equaGame = TRUE,
  correction = NA,
  transfo = NULL,
  data_weights = rep(1, nrow(database)),
  residuals = FALSE,
  measure,
  theta = NULL
)
}
\arguments{
\item{database}{A data.frame with data used to estimate the model.}

\item{model_eco}{An econometric model.}

\item{equation}{Either the "selection" equation or the "outcome" equation from a Tobit II model.}

\item{equaGame}{Whether the game is equalized or not (either TRUE, by default, or FALSE).}

\item{correction}{To keep the sum of the outcome constant across coalitions (either NA by default, "abs" or "rel").}

\item{transfo}{Specify a function used to transform the counterfactual outcome distribution before computing the inequality (default is NULL).}

\item{data_weights}{A vector of individuals weights (vector of 1 by default).}

\item{residuals}{Whether the residuals of the model are considered as an attribute in the decomposition or not (TRUE by default, or FALSE).}

\item{measure}{The function to use to compute the inequality, either Gini(),
Gini_G(), Entropy(), Atkinson(), Kolm(), Var().}

\item{theta}{The sensibility parameter value for the computation of the inequality (1 by default).}
}
\value{
Returns a list with: 1) The Shapley value for each attribute, 2) The pure marginal contribution of each attribute, 3) The pairwise interaction terms.
}
\description{
Compute the Shapley decomposition of the inequality of an outcome variable across different attributes from an econometric model.
}
\examples{
shapleyDecompo(database = exData,
               model_eco = exTobitModel,
               equation = "outcome",
               equaGame = TRUE,
               correction = NA,
               data_weights = exData$extridf,
               residuals = TRUE,
               transfo = exp,
               measure = Atkinson,
               theta = 1)
}
